web/requirements.txt
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
boto3==1.10.14
botocore==1.13.14
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
crayons==0.3.0
defusedxml==0.6.0
django-environ==0.4.5
docutils==0.15.2
flower==0.9.3
idna==2.8
importlib-metadata==0.21
jmespath==0.9.4
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.2.0
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tornado==5.1.1
urllib3==1.25.3
vine==1.3.0
watchtower==0.7.3
webdriver-manager==2.2.0
zipp==0.6.0
django-cors-headers
django-rest-framework-mongoengine
djangorestframework~=3.12.4
mongoengine~=0.23.1
rejson~=0.5.4
Django==2.2.13
gunicorn==19.9.0
psycopg2==2.7.7
